Added support for 1.21.3 and 1.21.4. Added SkyLib as a dependency. Updated README.md.
