+++ Core/src/main/java/com/github/lukesky19/skytrials/configuration/record/PlayerData.java
@@ -1,0 +1,25 @@
+/*
+    SkyTrials is a mob arena plugin inspired by the Minecraft 1.21 Trial Chambers using Trial Spawners and Vault blocks.
+    Copyright (C) 2024  lukeskywlker19
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as published
+    by the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+package com.github.lukesky19.skytrials.configuration.record;
+
+import com.github.lukesky19.skylib.libs.configurate.objectmapping.ConfigSerializable;
+
+import java.util.HashMap;
+
+@ConfigSerializable
+public record PlayerData(String configVersion, HashMap<String, Long> cooldowns) {}
