+++ Core/src/main/java/com/github/lukesky19/skytrials/data/Trial.java
