+++ Core/src/main/java/com/github/lukesky19/skytrials/listener/LogoutListener.java
