+++ Core/src/main/resources/plugin.yml
