+++ Core/src/main/resources/settings.yml
