+++ Core/src/main/resources/trials/example.yml
