+++ README.md
@@ -13,30 +13,36 @@
 * WorldGuard
 * PlaceholderAPI
 
+## Issues, Bugs, or Suggestions
+* Please create a new [GitHub Issue](https://github.com/lukesky19/SkyTrials/issues) with your issue, bug, or suggestion.
+* If an issue or bug, please post any relevant logs containing errors related to SkyWelcome and your configuration files.
+* I will attempt to solve any issues or implement features to the best of my ability.
+
 ## FAQ
 Q: What versions does this plugin support?
 
-A: 1.21.1.
+A: 1.21.1, 1.21.3, and 1.21.4
 
 Q: Are there any plans to support any other versions?
 
-A: No.
+A: I will always support newer versions of the game. I have no plans to support any version older than 1.21.1 at this time.
 
 Q: Does this work on Spigot and Paper?
 
-A: This will work on Paper and potentially forks of Paper (Untested).
+A: This plugin only works with Paper, it makes use of many newer API features that don't exist in the Spigot API. There are no plans to support Spigot.
 
 Q: Is Folia supported?
 
 A: There is no Folia support at this time. I may look into it in the future though.
 
-## Issues or Bugs
-* Create a new GitHub Issue describing your issue.
-* Please post any relevant logs containing errors related to SkyTrials and your configuration files.
-* I will attempt to solve any issues to the best of my ability.
+## For Server Admins/Owners
+* Download the plugin [SkyLib](https://github.com/lukesky19/SkyLib/releases).
+* Download the plugin from the releases tab and add it to your server.
 
 ## Building
-```./gradlew build```
+* Go to [SkyLib](https://github.com/lukesky19/SkyLib) and follow the "For Developers" instructions.
+* Then run:
+  ```./gradlew build```
 
 ## Why AGPL3?
 I wanted a license that will keep my code open source. I believe in open source software and in-case this project goes unmaintained by me, I want it to live on through the work of others. And I want that work to remain open source to prevent a time when a fork can never be continued (i.e., closed-sourced and abandoned).
