+++ settings.gradle.kts
@@ -1,1 +1,11 @@
+include(":Core")
+include(":1_21_1")
+include(":1_21_3")
+
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+        maven("https://repo.papermc.io/repository/maven-public/")
+    }
+}
